Running both localization algorithms programmatically
Analysis:
Setting each algorithm up requires changing the function call
and filenames in more than one location.

Resolution:
Setting up the algorithms so that they can be called programmatically
using dictionaries of parameters. New algorithms would need to contain
the "ProjectionFT" and "FieldOfView" parameters.

Note that the cells that graph the results still only look at one of the
algorithm's outputs. And the last cell is still hard-coded with results.

Ideally all the results would be stored in a pandas data frame instead
of being copied around in files and stored in multi-dimensional arrays.